create share button and share overlay with link can be copied to clipboard, some bugs remain. save before add loading animation for data fetch
