+++ src/database-demo/DatabaseDemo.css
@@ -215,12 +215,12 @@ rect:hover{
 .copy-success-message {
     font-size: 10pt;
     position: fixed;
-    bottom: 12px;
+    bottom: 0px;
     left: 50%;
     transform: translateX(-50%);
     background-color: rgba(0, 0, 0, 0.4);
     color: white;
-    padding: 10px 20px;
+    padding: 1%;
     border-radius: 5px;
     z-index: 1001;
     box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
