--- src/database-demo/d3-attom-demo/json-sample-500-mohawk.json
@@ -1,210 +1,0 @@
-{
-	"status": {
-		"version": "1.0.0",
-		"code": 0,
-		"msg": "SuccessWithResult",
-		"total": 1,
-		"page": 1,
-		"pagesize": 10,
-		"transactionID": "0c57aa54c61badc85a192c7e6dfe7006",
-		"attomId": 184702395
-	},
-	"property": [
-		{
-			"identifier": {
-				"Id": 184702395,
-				"fips": "08013",
-				"apn": "1577-04-2-15-020",
-				"attomId": 184702395
-			},
-			"area": {
-				"countrySecSubd": "Boulder"
-			},
-			"address": {
-				"country": "US",
-				"countrySubd": "CO",
-				"line1": "500 MOHAWK DR APT 305",
-				"line2": "BOULDER, CO 80303",
-				"locality": "BOULDER",
-				"matchCode": "ExaStr",
-				"oneLine": "500 MOHAWK DR APT 305, BOULDER, CO 80303",
-				"postal1": "80303",
-				"postal2": "3751",
-				"postal3": "C006"
-			},
-			"location": {
-				"accuracy": "Rooftop",
-				"latitude": "39.996795",
-				"longitude": "-105.235725",
-				"distance": 0.0,
-				"geoid": "CO08013, CS0890342, DB0802490, ND0004845235, PL0807850, RS0004818773, SB0000075213, SB0000075220, SB0000075222, ZI80303",
-				"geoIdV4": {
-					"CO": "4c746922ad37c5e0e9c397a1d56cbd50",
-					"CS": "cb51f87bd02deb9c049760af61dcc989",
-					"DB": "082f7411669bda9b2e9c5e1a61a610ec",
-					"N1": "9668dca0215d4d1a65734fd9667ca42d",
-					"N4": "cca6be7378576e6cbcc48dec39833ffb",
-					"PL": "ff044fd95d9be26fb3f2c6f3597fe5e5",
-					"SB": "c65df90325bcc6e6c53efbe5c3a5bb08, 75085900c9a76488f59a0802c561e8a9, 8615e14c3f8c250278c3ad643ef442ea",
-					"ZI": "a5308ec1377a7ca92500ba29ecbce3c5"
-				}
-			},
-			"salehistory": [
-				{
-					"sequence": 1,
-					"saleSearchDate": "2021-12-13",
-					"saleTransDate": "2021-12-13",
-					"transactionIdent": 979673022,
-					"buyerName": "YIQI NI,",
-					"amount": {
-						"saleDisclosureType": 0,
-						"saleDocType": "MORTGAGE",
-						"saleTransType": "Stand Alone Finance"
-					},
-					"mortgage": {
-						"FirstConcurrent": {
-							"trustDeedDocumentNumber": "0003934104",
-							"amount": 262800.0,
-							"lenderLastName": "ELEVATIONS CU"
-						},
-						"SecondConcurrent": {}
-					},
-					"vintage": {
-						"lastModified": "2022-07-26"
-					}
-				},
-				{
-					"sequence": 2,
-					"saleSearchDate": "2020-05-13",
-					"saleTransDate": "2020-05-03",
-					"transactionIdent": 876245692,
-					"buyerName": "YIQI NI,",
-					"sellerName": "ZHONGXIAN HU",
-					"amount": {
-						"saleRecDate": "2020-05-13",
-						"saleDisclosureType": 0,
-						"saleDocType": "DEED",
-						"saleDocNum": "0003783774",
-						"saleTransType": "Nominal/Quit Claim"
-					},
-					"mortgage": {
-						"FirstConcurrent": {},
-						"SecondConcurrent": {}
-					},
-					"vintage": {
-						"lastModified": "2022-04-28"
-					}
-				},
-				{
-					"sequence": 3,
-					"saleSearchDate": "2014-01-30",
-					"saleTransDate": "2013-12-11",
-					"transactionIdent": 408932358,
-					"buyerName": "YIQI NI,",
-					"sellerName": "NI,YIQI",
-					"amount": {
-						"saleRecDate": "2014-01-30",
-						"saleDisclosureType": 0,
-						"saleDocType": "DEED",
-						"saleDocNum": "0003364229",
-						"saleTransType": "Nominal/Quit Claim"
-					},
-					"mortgage": {
-						"FirstConcurrent": {},
-						"SecondConcurrent": {}
-					},
-					"vintage": {
-						"lastModified": "2022-04-20"
-					}
-				},
-				{
-					"sequence": 4,
-					"saleSearchDate": "2013-12-11",
-					"saleTransDate": "2013-12-11",
-					"transactionIdent": 382786220,
-					"buyerName": "YIQI NI,",
-					"sellerName": "HELART FAMILY ASSET MGMT LLC",
-					"amount": {
-						"saleAmt": 195000,
-						"saleCode": "ESTIMATED Sale price is not available due to non-disclosure counties",
-						"saleRecDate": "2013-12-11",
-						"saleDisclosureType": 0,
-						"saleDocType": "DEED",
-						"saleDocNum": "0003356827",
-						"saleTransType": "Resale"
-					},
-					"mortgage": {
-						"FirstConcurrent": {},
-						"SecondConcurrent": {}
-					},
-					"vintage": {
-						"lastModified": "2018-05-13"
-					}
-				},
-				{
-					"sequence": 5,
-					"saleSearchDate": "2011-09-19",
-					"saleTransDate": "2011-08-19",
-					"transactionIdent": 315667610,
-					"buyerName": "HELART FAMILY ASSET MGMT LLC,",
-					"sellerName": "HELART,KEITH E",
-					"amount": {
-						"saleRecDate": "2011-09-19",
-						"saleDisclosureType": 0,
-						"saleDocType": "DEED",
-						"saleDocNum": "0003171599",
-						"saleTransType": "Nominal/Quit Claim"
-					},
-					"mortgage": {
-						"FirstConcurrent": {},
-						"SecondConcurrent": {}
-					},
-					"vintage": {
-						"lastModified": "2022-04-22"
-					}
-				},
-				{
-					"sequence": 6,
-					"saleSearchDate": "2005-03-08",
-					"saleTransDate": "2005-03-08",
-					"transactionIdent": 328340928,
-					"buyerName": "OWNER NAME UNAVAILABLE,",
-					"sellerName": "OWNER NAME UNAVAILABLE",
-					"amount": {
-						"saleRecDate": "2005-03-08",
-						"saleDisclosureType": 0,
-						"saleDocType": "DEED",
-						"saleDocNum": "0002684072",
-						"saleTransType": "Resale"
-					},
-					"mortgage": {
-						"FirstConcurrent": {},
-						"SecondConcurrent": {}
-					},
-					"vintage": {
-						"lastModified": "2018-05-12"
-					}
-				}
-			],
-			"owner": {
-				"corporateIndicator": "N",
-				"type": "INDIVIDUAL",
-				"description": "INDIVIDUAL",
-				"owner1": {
-					"fullName": "YIQI NI",
-					"lastName": "NI ",
-					"firstNameAndMi": "YIQI "
-				},
-				"owner2": {},
-				"owner3": {},
-				"owner4": {},
-				"absenteeOwnerStatus": "A",
-				"mailingAddressOneline": "22 DELL ST, SOMERVILLE, MA 02145-4122"
-			},
-			"vintage": {
-				"lastModified": "2022-04-22",
-				"pubDate": "2022-04-22"
-			}
-		}
-	]
-}
